Feat endpoint get admin account

+++ src/controller/accountController.ts
@@ -53,4 +53,13 @@ const deleteById = async (req: Request, res: Response, next: NextFunction) => {
   }
 };
 
-export { create, list, getById, updateById, deleteById };
+const listAdmin = async (req: Request, res: Response, next: NextFunction) => {
+  try {
+    const accounts = await accountData.listAdmin();
+    return res.status(200).json(responseBody("OK", null, accounts));
+  } catch (error) {
+    next(error);
+  }
+};
+
+export { create, list, getById, updateById, deleteById, listAdmin };
